Tests/LibWeb/Text/expected/css/calc-round-strategy-roundtrip.txt

round(2rem, 5px) => round(nearest, 2rem, 5px) => 30px
round(nearest, 2rem, 5px) => round(nearest, 2rem, 5px) => 30px
round(up, 2rem, 5px) => round(up, 2rem, 5px) => 35px
round(down, 2rem, 5px) => round(down, 2rem, 5px) => 30px
round(to-zero, 2rem, 5px) => round(to-zero, 2rem, 5px) => 30px
